fix: preserve exact counter reads across core, bindings and examples (F3) (#73)
* fix: read counters exactly past the 64-bit boundary (F3)

The Lean model reads a G-Counter as an unbounded natural
(`Crdt.gcounterValue : ℕ`, crdt-lean Crdt/Instances.lean:95) and a
PN-Counter as an unbounded integer (`Crdt.pncounterValue : ℤ`,
Crdt/Instances.lean:147). The crate summed u64 tallies into a u64, which
wraps in release and panics in debug, and cast both sides to i64, so a
single tally of 2^63 read as i64::MIN. The comment on the PN read admitted
the divergence ("ℤ in the model, i64 here") instead of closing it.

`GCounter::value` now returns `u128` and `PnCounter::value` returns `i128`.
Both are exact for every representable state: at most isize::MAX
coordinates of at most u64::MAX sum to strictly below 2^127, so the wider
read never wraps and never saturates. The per-replica u64 state, merge
(pointwise max, closed over u64) and the wire format are unchanged.

Bindings: Python returns an exact `int`; WASM returns an exact `bigint`
built from the decimal total, with the TypeScript signature unchanged;
the C ABI replaces `safemesh_gcounter_value` with
`safemesh_gcounter_try_value(counter, out)`, which returns the new
`ValueOverflow` status and leaves `out` untouched rather than writing a
wrapped total. The header is regenerated with cbindgen and the header pin
test carries that output.

Planted tests in core_api.rs and ffi.rs assert on the number a caller
reads at the boundary for both halves (the sum and the cast); the
planted lossy bodies turn them red in both profiles.


* fix: propagate counter read errors in examples and reject rounded Lab reads

---------

+++ rust/crates/safemesh-crdt/examples/break_it.rs
@@ -60,7 +60,7 @@ struct Packet {
 
 #[derive(Clone, Debug, PartialEq, Eq)]
 struct Projection {
-    counter: u64,
+    counter: u128,
     supplies: Vec<u64>,
     text_positions: Vec<u64>,
 }
